R6 Class representing a batch container.

R6 Class representing a batch container.

R6 Class representing a batch container.

Methods

Public methods

BatchContainer$new()

BatchContainer$has_samples()

BatchContainer$get_samples()

BatchContainer$distribute_samples()

BatchContainer$print()

BatchContainer$clone()

Method new()

Usage

BatchContainer$new(
  dimensions,
  interactions = FALSE,
  interaction_weights = NULL,
  exclude = NULL
)

Method has_samples()

Returns TRUE if batch has samples.

Usage

BatchContainer$has_samples()

Method get_samples()

Return table with samples and sample assignment.

Usage

BatchContainer$get_samples(
  assignment = TRUE,
  include_id = FALSE,
  remove_empty_locations = FALSE
)

Arguments

assignment

Return sample assignment. If FALSE, only samples table is returned, with out batch assignment.

include_id

Keep .sample_id in the tibble.

remove_empty_locations

Removes empty locations from the result tibble.

Returns

tibble with samples and sample assignment. General method for performing sample assignment.

Method distribute_samples()

Usage

BatchContainer$distribute_samples(
  samples = NULL,
  distribution_function = distribute_random,
  distribution_function_args = list(),
  random_seed = NULL
)

Arguments

samples

data.frame with samples and unique .sample_id field.

random_seed

If set will fix random seed and then return the original value

ditribution_function

Function performing sample distribution

ditribution_function_args

Arguments for the distribution function

Returns

Any extra results provided by the distribution function. E.g., optimization trajectory or optimization score.

Method print()

Usage

BatchContainer$print(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

BatchContainer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.