get Score

Summary score for an experimental layout based on the distribution of levels of the factors to be balanced.

getScore(
  layout,
  balance,
  sc_groups,
  sc_tests,
  bal_weights = rep(1, length(balance)),
  sc_weights = rep(1, length(sc_groups))
)

Arguments

layout
a data.frame with the factors to be balanced and their current positions

balance
a vector with the names of experimental conditions to be balanced

sc_groups
list of dimension name groups to be used jointly for scoring

sc_tests
list of tests to use for each scoring group

bal_weights
named vector of weights for the factors to be balanced (default all 1)

sc_weights
named vector of weights for the dimensions (default all 1)

Value

the summarized penalty score